Separate some geoserver assert methods into geoserver_proxy
