+++ tests/asserts/final/publication/__init__.py
@@ -1,6 +1,6 @@
 from tests import Action
 from test_tools import process_client
-from . import geoserver, internal, internal_rest, rest, metadata
+from . import geoserver, geoserver_proxy, internal, internal_rest, rest, metadata
 
 
 IS_LAYER_COMPLETE_AND_CONSISTENT = [
@@ -23,8 +23,8 @@ IS_LAYER_COMPLETE_AND_CONSISTENT = [
     Action(internal.stats_preserved_in_normalized_raster, dict()),
     Action(rest.is_complete_in_rest, dict()),
     Action(rest.mandatory_keys_in_rest, dict()),
-    Action(geoserver.is_complete_in_workspace_wms_1_3_0, dict()),
-    Action(geoserver.workspace_wfs_2_0_0_capabilities_available_if_vector, dict()),
+    Action(geoserver_proxy.is_complete_in_workspace_wms_1_3_0, dict()),
+    Action(geoserver_proxy.workspace_wfs_2_0_0_capabilities_available_if_vector, dict()),
 ]
 
 IS_MAP_COMPLETE_AND_CONSISTENT = [
