+++ tests/asserts/final/publication/geoserver.py
@@ -1,43 +1,12 @@
 import copy
 import math
 from urllib import parse
-import requests
 
 import crs as crs_def
 from layman import app, settings, util as layman_util
 from layman.common import bbox as bbox_util
-from layman.layer.geoserver import wfs, wms, util as gs_util
-from test_tools import util as test_util, geoserver_client, process_client, assert_util
-
-
-def is_complete_in_workspace_wms_1_3_0(workspace, publ_type, name, headers):
-    assert publ_type == process_client.LAYER_TYPE
-
-    version = '1.3.0'
-    with app.app_context():
-        wms_url = test_util.url_for('geoserver_proxy_bp.proxy', subpath=workspace + settings.LAYMAN_GS_WMS_WORKSPACE_POSTFIX + '/ows')
-    wms_inst = gs_util.wms_proxy(wms_url, version=version, headers=headers)
-    assert wms_inst.contents
-    assert name in wms_inst.contents
-    wms_layer = wms_inst.contents[name]
-    for style_name, style_values in wms_layer.styles.items():
-        assert 'legend' in style_values, f'style_name={style_name}, style_values={style_values}'
-
-
-def workspace_wfs_2_0_0_capabilities_available_if_vector(workspace, publ_type, name):
-    with app.app_context():
-        internal_wfs_url = test_util.url_for('geoserver_proxy_bp.proxy', subpath=workspace + '/wfs')
-
-    with app.app_context():
-        file_info = layman_util.get_publication_info(workspace, publ_type, name, {'keys': ['file']})['file']
-    file_type = file_info['file_type']
-    if file_type == settings.FILE_TYPE_VECTOR:
-        r_wfs = requests.get(internal_wfs_url, params={
-            'service': 'WFS',
-            'request': 'GetCapabilities',
-            'version': '2.0.0',
-        })
-        assert r_wfs.status_code == 200
+from layman.layer.geoserver import wfs, wms
+from test_tools import geoserver_client, process_client, assert_util
 
 
 def feature_spatial_precision(workspace, publ_type, name, *, feature_id, crs, exp_coordinates, precision):
